club/serializers.py
from rest_framework import serializers
from django.contrib.auth.models import User
from .models import Club, Membership


class ClubSerializer(serializers.ModelSerializer):
    class Meta:
        model = Club
        fields = '__all__'
        
        
        
class MembershipSerializer(serializers.ModelSerializer):
    class Meta:
        model = Membership
        fields = '__all__'